requirement.txt
Django==2.2.2
djangorestframework==3.10.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.8.0
drf-writable-nested==0.6.3
Pillow==7.2.0
PyJWT==1.7.1
python-dotenv==0.14.0
pytz==2022.1
sqlparse==0.4.2
